refactor(download): use tailwind in download component
