--- src/components/download/style.less
@@ -1,28 +1,0 @@
-/*
-License:
-
-This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of
-the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
-*/
-
-.main {
-  padding: 1rem;
-  padding-top: 0;
-  margin: 1rem auto;
-  text-align: center;
-  background-color: #d3d3d3;
-  border-radius: 5px;
-  color: #222;
-}
-
-.main.todo {
-  padding-top: 1rem;
-}
-
-.link {
-  display: inline-block;
-}
-
-.link + .link {
-  margin-left: 0.5rem;
-}
